#57 style : ui 모델 package 빼기
