+++ feature/color/src/main/kotlin/com/titi/app/feature/color/model/ColorUiState.kt
@@ -1,0 +1,7 @@
+package com.titi.app.feature.color.model
+
+import com.airbnb.mvrx.MavericksState
+
+data class ColorUiState(
+    val colors: List<Long> = emptyList(),
+) : MavericksState
