+++ feature/main/src/main/kotlin/com/titi/app/feature/main/model/SplashResultState.kt
@@ -1,14 +1,17 @@
-package com.titi.app.feature.time
+package com.titi.app.feature.main.model
 
-import android.os.Parcelable
 import com.titi.app.doamin.daily.model.Daily
 import com.titi.app.domain.color.model.TimeColor
 import com.titi.app.domain.time.model.RecordTimes
-import kotlinx.parcelize.Parcelize
 
-@Parcelize
 data class SplashResultState(
     val recordTimes: RecordTimes = RecordTimes(),
     val timeColor: TimeColor = TimeColor(),
     val daily: Daily? = null,
-) : Parcelable
+)
+
+fun SplashResultState.toFeatureTimeModel() = com.titi.app.feature.time.model.SplashResultState(
+    recordTimes = recordTimes,
+    timeColor = timeColor,
+    daily = daily,
+)
