+++ feature/measure/src/main/kotlin/com/titi/app/feature/measure/model/SplashResultState.kt
@@ -1,17 +1,14 @@
-package com.titi.app.feature.main.ui
+package com.titi.app.feature.measure.model
 
+import android.os.Parcelable
 import com.titi.app.doamin.daily.model.Daily
 import com.titi.app.domain.color.model.TimeColor
 import com.titi.app.domain.time.model.RecordTimes
+import kotlinx.parcelize.Parcelize
 
+@Parcelize
 data class SplashResultState(
     val recordTimes: RecordTimes = RecordTimes(),
     val timeColor: TimeColor = TimeColor(),
     val daily: Daily? = null,
-)
-
-fun SplashResultState.toFeatureTimeModel() = com.titi.app.feature.time.SplashResultState(
-    recordTimes = recordTimes,
-    timeColor = timeColor,
-    daily = daily,
-)
+) : Parcelable
