+++ feature/time/src/main/kotlin/com/titi/app/feature/time/ui/timer/TimerScreen.kt
@@ -29,7 +29,6 @@ import com.titi.app.core.designsystem.R
 import com.titi.app.core.designsystem.component.TdsTimer
 import com.titi.app.core.designsystem.theme.TdsColor
 import com.titi.app.core.util.toJson
-import com.titi.app.feature.time.SplashResultState
 import com.titi.app.feature.time.content.TimeButtonContent
 import com.titi.app.feature.time.content.TimeCheckDailyDialog
 import com.titi.app.feature.time.content.TimeColorDialog
@@ -37,6 +36,8 @@ import com.titi.app.feature.time.content.TimeDailyDialog
 import com.titi.app.feature.time.content.TimeHeaderContent
 import com.titi.app.feature.time.content.TimeTaskContent
 import com.titi.app.feature.time.content.TimeTimerDialog
+import com.titi.app.feature.time.model.SplashResultState
+import com.titi.app.feature.time.model.TimerUiState
 import com.titi.app.feature.time.ui.task.TaskBottomSheet
 import org.threeten.bp.ZoneOffset
 import org.threeten.bp.ZonedDateTime
